change to download init.d scripts
Some testing at Picademy revealed an issue with these instructions. It was to do with copying and pasting the init.d scripts. It took me a while to find out why. It was an invalid space character being introduced by the way the forum displays the code inside a code-block on the Midori browser. So I've fixed it by providing the raw files that you just download using wget. Instructions have been updated to include this.