Copy edit
In the cloning SD card section you mention the Win32 Disk Imager, but it's not mentioned anywhere else; earlier the user is told to install Etcher. I've removed the name in the cloning section, but the other steps might need changing, depending on whether you want the user to use Etcher or Win32.